backend/requirements.txt
fastapi
uvicorn[standard]
psycopg[pool]
python-dotenv
